Create ability to smartly wait for outfiles in jobcombine.sh with --wait
If you have multiple hosts connected through nfsjobtrain, there's a good
chance that every host will exectute the job at a different time. This
means that if you want to get all the results from the hosts, you'll
have to wait up to the cronjob time set. In practice, the time required
for waiting is often much less than the crontab time set, but it's
impossible to know this ahead of time.

This commit gets around this by adding a "-w/--wait TIME" flag in
jobcombine.sh that causes the script to wait up till the specified time,
or until it finds all the outfiles before then.

Pretty neat! :^)